Test cases for `Kernel.linop`
